backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy
redis==5.0.0

# Background job queue
rq==1.13.0
